Refactorings and fixings
